requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

torch==2.2.2+cpu
torchvision==0.17.2+cpu

transformers
fastapi
uvicorn
pillow
opencv-python-headless
scikit-learn
mlflow
python-multipart